Move detecting irregular whitespace to a markdown parser and add tests
