+++ tests/authoring/authoring.fsproj
@@ -58,4 +58,8 @@
     <Compile Include="Directives\IncludeBlocks.fs" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="Linters\WhiteSpaceNormalizers.fs" />
+  </ItemGroup>
+
 </Project>
